+++ resources/js/fe-sky-app/app-router/appRouter.js
@@ -1,0 +1,7 @@
+import {createRouter, createWebHistory} from 'vue-router';
+import {routerConfig} from './routerConfig';
+
+export const appRouter = createRouter({
+  history: createWebHistory(),
+  routes: routerConfig
+});
