+++ resources/js/fe-sky-app/app-router/index.js
@@ -1,1 +1,1 @@
-export * from './router-config';
+export {appRouter} from './appRouter';
