+++ resources/js/fe-sky-app/modules/auth/index.js
@@ -1,0 +1,9 @@
+export {
+  setToken,
+  getToken,
+  removeToken,
+  isLogin
+} from './storage-token';
+export {
+  refreshCurrentUser
+} from './actions/refreshCurrentUser';
