+++ resources/js/fe-sky-app/modules/auth/storage-token.js
@@ -1,0 +1,11 @@
+export const setToken = (token) => {
+  localStorage.setItem('token', token);
+};
+
+export const getToken = () => localStorage.getItem('token');
+
+export const removeToken = () => {
+  localStorage.removeItem('token');
+};
+
+export const isLogin = () => Boolean(getToken());
